refactor(jazz-tools): Phase 3 — split storage/mod.rs along trait + impl seams
storage/mod.rs was ~8K LOC mixing three concerns. Split into three files:

  - mod.rs            (~2.7K) — trait-adjacent types/codecs and the value
                                encoder helpers (is_double_zero, encode_value)
                                used by both the trait surface and key_codec
  - storage_trait.rs  (~2.1K) — the Storage trait + the Box<T: Storage>
                                forwarding impl; pulls in trait-adjacent
                                types via `use super::*;`
  - memory.rs         (~3.2K) — MemoryStorage struct, helpers, and the
                                pre-existing test module that primarily
                                exercises MemoryStorage against the trait

mod.rs re-exports `Storage` and `MemoryStorage` so every external import
(`storage::Storage`, `storage::MemoryStorage`) keeps resolving unchanged.

The opfs_btree single-file → opfs_btree/{native,wasm,mod}.rs cfg-fork
restructure listed in the spec is deferred — it's cosmetic relative to
this split and warrants its own PR. Mentioned in the spec as part of
Phase 3 but factoring it cleanly while keeping the shared `AnyFile`
dispatch enum is non-trivial; tracking it as Phase 3.5.

Spec: specs/todo/c_later/crate_structure_cleanup.md (Phase 3).
